Fixed phantomjs-testrunner.js to work on Windows
The line:
window.fs_path_separator = "\%fs_path_separator%";

yields a syntax error on windows. %fs_path_separator% is \, and so the resulting line is:
window.fs_path_separator = "\";

But this is not valid JS (backslash escapes the closing quote).

This should solve issue #28


Prefixing with a backslash solves the problem on Windows, and is ignored on Mac/Unix, so everything still works there.